Modified Android's Game class to function more like XNA's Game class in that Game.Initialize calls Initialize on all GameComponents and then calls Game.LoadContent.  Game.Update and Game.Draw no longer check for "initialized".
The default ContentManager is created in the Game constructor instead of on first access.

Added a public static property called Activity (gets and sets _contextInstance).  This is Android-specific and avoids having to pass the Activity as a parameter to the Game constructor.  This means that we remove all of the #if ANDROID ... #else ... #endif in all projects around the constructors of their Game classes.  The Game.Activity static property is set in the Activity.OnCreate method before constructing the game instance.